+++ examples/beforeEnter/src/vite-env.d.ts
@@ -1,0 +1,7 @@
+/// <reference types="vite/client" />
+interface Window {
+  __INITIAL_DATA__: {
+    basename: string;
+    mode: string;
+  };
+}